fix: CRITICAL RESCUE - Fix Kernel Crash & Improve Auth UX
GAD-000 LEVEL VIOLATION FIX: System was crashing silently with ModuleNotFoundError
while UI showed "Guest Mode" with no transparency.

CHANGES:
1. **Fixed Missing Module (vibe_core/ledger.py)**
   - Extracted SQLiteLedger and InMemoryLedger to separate ledger.py module
   - Fixes: ModuleNotFoundError: No module named 'vibe_core.ledger'
   - Provides: Hash-chained persistent event ledger with tamper detection

2. **Updated Kernel Imports (vibe_core/kernel_impl.py)**
   - Now imports ledger implementations from vibe_core.ledger
   - Removed duplicate ledger class definitions (~370 lines)
   - Maintains backward compatibility

3. **Simplified Launcher (vibe_launcher.py)**
   - Removed incorrect VibeLedger instantiation in boot_kernel()
   - Kernel creates its own ledger instance - no duplication needed
   - Fixed kernel.agents -> kernel.agent_registry reference

4. **Enhanced Frontend UI (gateway/static/index.html)**
   - UNLOCKED: Command input for Guest Mode (zero-click entry)
   - REMOVED: Forced authentication modal on startup
   - ADDED: Error overlay for kernel crashes with retry mechanism
   - IMPROVED: WebSocket error handlers show transparent failure messages

ARCHITECTURE CHANGE (Vimana Style):
- Authentication is now server-side (Launcher detects keys at boot)
- Frontend is pure observer (no private keys in browser)
- Guest Mode is fully functional (read-only via Milk Ocean Router)
- Errors are now transparent (error overlay shows traceback)

TESTED: System boots without crash, Input is enabled, Errors are visible
